rename: rename AppProviders
